// src/app/sistema/administrativo/cadastro-de-colaborador/enums/setor.ts
export enum Setor {
  FRONTEND = 'FRONTEND',
  PROFESSOR_UFC = 'PROFESSOR_UFC',
  MONITOR = 'MONITOR',
  BACKEND = 'BACKEND',
  FULLSTACK = 'FULLSTACK',
  DEVOPS = 'DEVOPS',
  QA = 'QA',
  PROCESSOS = 'PROCESSOS',
  REQUISITOS = 'REQUISITOS',
  UI_UX_DESIGN = 'UI/UX DESIGN',
  CONSULTOR = 'CONSULTOR',
}